Roads: add new road generation algorithm
This algorithm is based on "Road Tool" script which is under free
license. It builds roads using heightmap/alphamap data of terrain
and creates the corresponding mesh. Unfortunately, it has poor
performance and should be improved as soon as possible (seems,
the easiest way is to remove heightmap processing in it as we
don't use this feature so far <flat map>)